requirements.txt
# 核心依赖
librosa>=0.10.0
numpy>=1.21.0
pyaudio>=0.2.11
pyserial>=3.5
python-dotenv>=1.0.0
requests>=2.28.0

# AI 对话（DeepSeek API）
openai>=1.0.0

# 开发依赖
pytest>=7.0.0
